Title: 달에서 무게 구하기, Time: 96ms, Memory: 9MB, Status: Passed - Codetree

+++ 260128/달에서 무게 구하기/weight-on-the-moon.java
@@ -1,9 +1,9 @@
 public class Main {
     public static void main(String[] args) {
         // Please write your code here.
-        int weight=13;
-        int gravity=0.165;
+        double weight=13;
+        double gravity=0.165;
 
-        System.out.printf("%d * %d = %d",weight,gravity,weight*gravity);
+        System.out.printf("%.0f * %f = %f",weight,gravity,weight*gravity);
     }
 }
